Allow config to specify min/max
This is for cases when the chart wants a y-axis min smaller than, or a max larger than, the bounds of the series.

For example, it's quite common to want the chart to have a 0 value on the y-axis, regardless of the data. Also being able to specify the max gives the caller a chance to create a range that has 'sensible' rounded tick values on the y-axis.